feat: fold sign-in into index, add welcome toast, eliminate SignInScreen (#82)
Collapses the redundant auth entry point: index.tsx now hosts the full
login form, SignInScreen is deleted, and a 'Welcome back, {firstName}!'
toast fires on HomePage after successful login via react-native-toast-message.
Also tightens the RootLayoutNav spinner guard so unauthenticated cold
launches show the login form immediately instead of a blank spinner frame.
