+++ utils/__tests__/loginFlag.test.ts
@@ -1,0 +1,23 @@
+import { setWelcomeUser, consumeWelcomeUser } from "../loginFlag";
+
+describe("loginFlag", () => {
+  beforeEach(() => {
+    // Reset state between tests by consuming any pending flag
+    consumeWelcomeUser();
+  });
+
+  it("setWelcomeUser stores the first name", () => {
+    setWelcomeUser("Jordan");
+    expect(consumeWelcomeUser()).toBe("Jordan");
+  });
+
+  it("consumeWelcomeUser clears the flag after reading", () => {
+    setWelcomeUser("Jordan");
+    consumeWelcomeUser();
+    expect(consumeWelcomeUser()).toBeNull();
+  });
+
+  it("consumeWelcomeUser returns null when no flag is set", () => {
+    expect(consumeWelcomeUser()).toBeNull();
+  });
+});
